pyproject.toml
[tool.poetry]
name = "pzl_food"
version = "0.1.0"
description = "Food managment"
authors = [""]
maintainers = [""]

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pydantic = "^2.5.2"
dynaconf = "^3.2.4"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
sqlalchemy = "^2.0.23"
sqlalchemy-utils = "^0.41.1"
starlette-graphene3 = "^0.6.0"
graphene-sqlalchemy = "3.0.0rc1"
python-dateutil = "^2.8.2"
backports-zoneinfo = "^0.2.1"
python-multipart = "^0.0.6"
yarl = "^1.9.4"
defusedxml = "^0.7.1"

[tool.poetry.dev-dependencies]
anyio = "^3.7.1"
httpx = "^0.26.0"
isort = "^5.13.2"
black = "^23.12.1"
flake8 = "^7.0.0"
pylint = "^3.0.3"
pytest = "^7.4.4"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.3"
coverage = "^7.4.0"
greenlet = "^3.0.3"
genbadge = {extras = ["tests"], version = "^1.1.1"}

[tool.poetry.group.lint.dependencies]
isort = "5.13.2"
black = "^23.12.1"
flake8 = "^7.0.0"
pylint = "^3.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py38']
